Converted 'unsigned long long' to uint64_t (stdint.h) Added percentages to statistics Minor bug fix to 'msgslot' handling.

+++ src/output-plugins/sagan-snort.c
@@ -39,6 +39,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <pthread.h>
+#include <stdint.h>
 
 #include "sagan.h"
 #include "version.h"
@@ -74,7 +75,7 @@ int devdebug;
 
 int  threaddbc;
 
-unsigned long long sigcid;
+uint64_t sigcid;
 
 struct rule_struct *rulestruct;
 
@@ -280,12 +281,12 @@ return(0);
 /* Get the last used CID and increment it */
 /******************************************/
 
-unsigned long long get_cid ( int sensor_sid, int dbtype ) { 
+uint64_t get_cid ( int sensor_sid, int dbtype ) { 
 
 char sqltmp[MAXSQL]; 
 char *sql;
 char *sqlout;
-unsigned long long t_cid; 
+uint64_t t_cid; 
 
 
 snprintf(sqltmp, sizeof(sqltmp), "SELECT last_cid from sensor where sid=%d and hostname='%s' and interface='%s' and filter='%s' and detail=%d", sensor_sid, sagan_hostname, sagan_interface, sagan_filter, sagan_detail);
@@ -370,14 +371,14 @@ return(t_sig_id);
 /* Insert into event table */
 /***************************/
 
-void insert_event (int t_sid,  unsigned long long t_cid,  int t_sig_sid,  int dbtype,  char *date,  char *time ) { 
+void insert_event (int t_sid,  uint64_t t_cid,  int t_sig_sid,  int dbtype,  char *date,  char *time ) { 
 
 char sqltmp[MAXSQL];
 char *sql;
 
 pthread_mutex_lock( &db_mutex );
 
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO event(sid, cid, signature, timestamp) VALUES ('%d', '%llu', '%d', '%s %s')", t_sid, t_cid, t_sig_sid, date, time );
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO event(sid, cid, signature, timestamp) VALUES ('%d', '%lu', '%d', '%s %s')", t_sid, t_cid, t_sig_sid, date, time );
 sql=sqltmp;
 
 pthread_mutex_unlock( &db_mutex );
@@ -391,7 +392,7 @@ db_query( dbtype, sql );
 /* Insert data into iphdr and tcphdr - most of this is bogus as we're not really TCP/IP */
 /****************************************************************************************/
 
-void insert_hdr(int t_sid, unsigned long long t_cid, char *t_ipsrc, char *t_ipdst, int t_ipproto, int endian, int dbtype, int dst_port, int src_port) { 
+void insert_hdr(int t_sid, uint64_t t_cid, char *t_ipsrc, char *t_ipdst, int t_ipproto, int endian, int dbtype, int dst_port, int src_port) { 
 
 char ipsrcbit[MAXHOST];
 char ipdstbit[MAXHOST];
@@ -406,14 +407,14 @@ snprintf(ipdstbit, sizeof(ipdstbit), "%s", ip2bit(t_ipdst, endian));
 
 /* 4 == IPv4 */
 
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO iphdr VALUES ( '%d', '%llu', '%s', '%s', '4', '0', '0', '0', '0', '0', '0', '0', '%d', '0' )", t_sid, t_cid, ipsrcbit, ipdstbit, t_ipproto );
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO iphdr VALUES ( '%d', '%lu', '%s', '%s', '4', '0', '0', '0', '0', '0', '0', '0', '%d', '0' )", t_sid, t_cid, ipsrcbit, ipdstbit, t_ipproto );
 
 sql=sqltmp;
 db_query( dbtype, sql );
 
 /* "tcp" */
 if ( t_ipproto == 6 )  {
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO tcphdr VALUES ( '%d', '%llu', '%d', '%d', '0', '0', '0', '0', '0', '0', '0', '0'  )", t_sid, t_cid, src_port, dst_port  );
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO tcphdr VALUES ( '%d', '%lu', '%d', '%d', '0', '0', '0', '0', '0', '0', '0', '0'  )", t_sid, t_cid, src_port, dst_port  );
 sql=sqltmp;
 db_query( dbtype, sql );
 } 
@@ -421,7 +422,7 @@ db_query( dbtype, sql );
 /* "udp" */
 
 if ( t_ipproto == 17 )  {
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO udphdr VALUES ( '%d', '%llu', '%d', '%d', '0', '0' )", t_sid, t_cid, src_port, dst_port  );
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO udphdr VALUES ( '%d', '%lu', '%d', '%d', '0', '0' )", t_sid, t_cid, src_port, dst_port  );
 sql=sqltmp;
 db_query( dbtype, sql );
 }
@@ -430,7 +431,7 @@ db_query( dbtype, sql );
 /* May expand on this if there's actually a use for it */
 
 if ( t_ipproto == 1 ) { 
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO icmphdr VALUES ( '%d', '%llu', '8', '8', '0', '0', '0' )", t_sid, t_cid );
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO icmphdr VALUES ( '%d', '%lu', '8', '8', '0', '0', '0' )", t_sid, t_cid );
 sql=sqltmp;
 db_query( dbtype, sql );
 }
@@ -442,13 +443,13 @@ db_query( dbtype, sql );
 /* Insert into payload table */
 /*****************************/
 
-void insert_payload ( int t_sid, unsigned long long t_cid, char *t_hex_data,  int dbtype ) { 
+void insert_payload ( int t_sid, uint64_t t_cid, char *t_hex_data,  int dbtype ) { 
 
 char sqltmp[MAXSQL]; 
 char *sql;
 
 pthread_mutex_lock( &db_mutex );
-snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO data(sid, cid, data_payload) VALUES ('%d', '%llu', '%s')", t_sid, t_cid, t_hex_data);
+snprintf(sqltmp, sizeof(sqltmp), "INSERT INTO data(sid, cid, data_payload) VALUES ('%d', '%lu', '%s')", t_sid, t_cid, t_hex_data);
 sql=sqltmp;
 pthread_mutex_unlock( &db_mutex );
 db_query( dbtype, sql );
@@ -464,7 +465,7 @@ void record_last_cid ( void )  {
 char sqltmp[MAXSQL];
 char *sql;
 
-snprintf(sqltmp, sizeof(sqltmp), "UPDATE sensor SET last_cid='%llu' where sid=%d and hostname='%s' and interface='%s' and filter='%s' and detail=%d", sigcid, sensor_id, sagan_hostname, sagan_interface, sagan_filter, sagan_detail);
+snprintf(sqltmp, sizeof(sqltmp), "UPDATE sensor SET last_cid='%lu' where sid=%d and hostname='%s' and interface='%s' and filter='%s' and detail=%d", sigcid, sensor_id, sagan_hostname, sagan_interface, sagan_filter, sagan_detail);
 sql=sqltmp;
 db_query( dbtype, sql );
